refactor(auth): move helper logic outside controller body
